refactor+feat: arc spacing, sign position, per-row shelf counts, SceneSignManager rename
Rename:
- CategorySignSystem.ts -> SceneSignManager.ts (class/type names updated throughout)
  Addresses PR#38 comment: too narrowly named for a generic sign placement system

Arc layout improvements:
- rowRadiusStep: 2.8 -> 4.0 (more breathing room between rings)
- shelvesPerRowByRow support: outer rings get proportionally more shelves
  (row 0: N, row 1: 1.2N, row 2: 1.4N, row 3: 1.6N, row 4: 1.8N)
- shelvesPerRowByRow TD comment + field added to ArcLayoutConfig with plan
  for inverted-layout (front sparse, back dense) per design discussion
- DEFAULTS type updated to handle optional shelvesPerRowByRow field

Recently Played sign:
- Position moved from Z=-1.5 to Z=-6.4 (between ring 0 at r=5 and ring 1 at r=7.8)
- Size increased from 3.0x0.5 to 4.0x0.65
